+++ test/shopify-cli/commands/connect_test.rb
@@ -7,64 +7,22 @@ module ShopifyCli
 
       def test_connect_asks_project_type_and_writes_yml_when_no_project_exists
         ShopifyCli::Project.stubs(:has_current?).returns(false)
-        Resources::EnvFile.expects(:parse_external_env).returns({})
-        ShopifyCli::PartnersAPI::Organizations.stubs(:fetch_with_app).returns(partners_api_response)
         CLI::UI::Prompt.expects(:ask).with(@context.message('core.connect.project_type_select')).returns('node')
-        CLI::UI::Prompt.expects(:ask).with(@context.message('core.tasks.ensure_env.organization_select')).returns(101)
-        CLI::UI::Prompt.expects(:ask).with(
-          @context.message('core.tasks.ensure_env.development_store_select')
-        ).returns('store.myshopify.com')
-
-        Resources::EnvFile.any_instance.expects(:write)
+        ShopifyCli::Tasks::EnsureEnv.expects(:call).with(@context, regenerate: true).returns(org_response)
         ShopifyCli::Project.expects(:write)
         run_cmd('connect')
       end
 
       def test_connect_doesnt_write_yml_when_current_project_exists
-        ShopifyCli::PartnersAPI::Organizations.stubs(:fetch_with_app).returns(partners_api_response)
-        CLI::UI::Prompt.expects(:ask).with(@context.message('core.tasks.ensure_env.organization_select')).returns(101)
-        CLI::UI::Prompt.expects(:ask).with(
-          @context.message('core.tasks.ensure_env.development_store_select')
-        ).returns('store.myshopify.com')
-
         CLI::UI::Prompt.expects(:ask).with(@context.message('core.connect.project_type_select')).never
-        Resources::EnvFile.any_instance.expects(:write)
+        ShopifyCli::Tasks::EnsureEnv.expects(:call).with(@context, regenerate: true).returns(org_response)
         ShopifyCli::Project.expects(:write).never
         run_cmd('connect')
       end
 
-      def test_connect_adds_to_env_file_if_already_exists
-        ShopifyCli::Project.stubs(:has_current?).returns(false)
-        Resources::EnvFile.expects(:parse_external_env).returns({
-          scopes: 'read_products',
-          extra: { "additional" => "things" },
-        })
-        ShopifyCli::PartnersAPI::Organizations.stubs(:fetch_with_app).returns(partners_api_response)
-        CLI::UI::Prompt.expects(:ask).with(@context.message('core.connect.project_type_select')).returns('node')
-        CLI::UI::Prompt.expects(:ask).with(@context.message('core.tasks.ensure_env.organization_select')).returns(101)
-        CLI::UI::Prompt.expects(:ask).with(
-          @context.message('core.tasks.ensure_env.development_store_select')
-        ).returns('store.myshopify.com')
-
-        Resources::EnvFile.expects(:new).with(
-          api_key: "apikey",
-          secret: 1234,
-          shop: "store.myshopify.com",
-          scopes: "read_products",
-          extra: { "additional" => "things" }
-        ).returns(stub(:write))
-        ShopifyCli::Project.expects(:write)
-        run_cmd('connect')
-      end
-
       def test_connect_outputs_warnings_if_already_connected
-        ShopifyCli::PartnersAPI::Organizations.stubs(:fetch_with_app).returns(partners_api_response)
         CLI::UI::Prompt.expects(:ask).with(@context.message('core.connect.project_type_select')).never
-        CLI::UI::Prompt.expects(:ask).with(@context.message('core.tasks.ensure_env.organization_select')).returns(101)
-        CLI::UI::Prompt.expects(:ask).with(
-          @context.message('core.tasks.ensure_env.development_store_select')
-        ).returns('store.myshopify.com')
-        Resources::EnvFile.any_instance.expects(:write)
+        ShopifyCli::Tasks::EnsureEnv.expects(:call).with(@context, regenerate: true).returns(org_response)
         ShopifyCli::Project.expects(:write).never
         ShopifyCli::Project.stubs(:current_project_type).returns(:rails)
 
@@ -77,21 +35,8 @@ module ShopifyCli
 
       private
 
-      def partners_api_response
-        [{
-          "id" => 100,
-          "businessName" => "one",
-          "stores" => [{
-            "shopDomain" => "store.myshopify.com",
-          }],
-          "apps" => [{
-            "title" => "app",
-            "apiKey" => "apikey",
-            "apiSecretKeys" => [{
-              "secret" => 1233,
-            }],
-          }],
-        }, {
+      def org_response
+        {
           "id" => 101,
           "businessName" => "two",
           "stores" => [
@@ -105,7 +50,7 @@ module ShopifyCli
               "secret" => 1234,
             }],
           }],
-        }]
+        }
       end
     end
   end