+++ test/shopify-cli/task/ensure_env_test.rb
@@ -15,14 +15,15 @@ module ShopifyCli
       end
 
       def test_create_new_app_if_none_available
-        ShopifyCli::PartnersAPI::Organizations.stubs(:fetch_with_app).returns([{
+        response = [{
           "id" => 421,
           "businessName" => "one",
           "stores" => [{
             "shopDomain" => "store.myshopify.com",
           }],
           "apps" => [],
-        }])
+        }]
+        ShopifyCli::PartnersAPI::Organizations.expects(:fetch_with_app).with(@context).returns(response)
         CLI::UI::Prompt.expects(:ask).with(@context.message('core.tasks.ensure_env.app_name')).returns('new app')
         CLI::UI::Prompt.expects(:ask).with(@context.message('core.tasks.ensure_env.app_type.select')).returns('public')
         ShopifyCli::Tasks::CreateApiClient.expects(:call).with(
@@ -31,12 +32,14 @@ module ShopifyCli
           title: 'new app',
           type: 'public',
         ).returns({
-          "apiKey" => "ljdlkajfaljf",
-          "apiSecretKeys" => [{ "secret" => "kldjakljjkj" }],
+          "apiKey" => 1235,
+          "apiSecretKeys" => [{ "secret" => 1234 }],
           "id" => "12345678",
         })
-        Resources::EnvFile.any_instance.stubs(:write)
-        EnsureEnv.call(@context)
+        env_file = Minitest::Mock.new
+        Resources::EnvFile.expects(:new).with(new_env_file_values).returns(env_file)
+        env_file.expect(:write, nil, [@context])
+        assert_equal response.first, EnsureEnv.call(@context)
       end
 
       def test_no_prompt_if_one_app_and_org
@@ -48,55 +51,99 @@ module ShopifyCli
           }],
           "apps" => [{
             "title" => "app",
-            "apiKey" => 1234,
+            "apiKey" => 1235,
             "apiSecretKeys" => [{
-              "secret" => 1233,
+              "secret" => 1234,
             }],
           }],
         }]
-        ShopifyCli::PartnersAPI::Organizations.stubs(:fetch_with_app).returns(response)
-        Resources::EnvFile.any_instance.stubs(:write)
-        EnsureEnv.call(@context)
+        ShopifyCli::PartnersAPI::Organizations.expects(:fetch_with_app).with(@context).returns(response)
+        env_file = Minitest::Mock.new
+        Resources::EnvFile.expects(:new).with(new_env_file_values).returns(env_file)
+        env_file.expect(:write, nil, [@context])
+        assert_equal response.first, EnsureEnv.call(@context)
       end
 
       def test_uses_default_values_for_env_file
-        Resources::EnvFile.expects(:parse_external_env).with.raises(Errno::ENOENT)
-        ShopifyCli::PartnersAPI::Organizations.stubs(:fetch_with_app).returns(partners_api_response)
-        CLI::UI::Prompt.expects(:ask).with(@context.message('core.tasks.ensure_env.organization_select')).returns(101)
-        CLI::UI::Prompt.expects(:ask).with(
-          @context.message('core.tasks.ensure_env.development_store_select')
-        ).returns('store.myshopify.com')
-
-        Resources::EnvFile.expects(:new).with(
-          api_key: 1235,
-          secret: 1234,
-          shop: "store.myshopify.com",
-          scopes: "write_products,write_customers,write_draft_orders",
-          extra: {}
-        ).returns(stub(:write))
-        EnsureEnv.call(@context)
+        Resources::EnvFile.expects(:parse_external_env).raises(Errno::ENOENT)
+        ShopifyCli::PartnersAPI::Organizations.expects(:fetch_with_app).with(@context).returns(partners_api_response)
+        expect_user_prompts
+        env_file = Minitest::Mock.new
+        Resources::EnvFile.expects(:new).with(new_env_file_values).returns(env_file)
+        env_file.expect(:write, nil, [@context])
+        assert_equal partners_api_response[1], EnsureEnv.call(@context)
       end
 
       def test_keep_existing_env_values
-        Resources::EnvFile.expects(:parse_external_env).with.returns({ host: "host" })
-        ShopifyCli::PartnersAPI::Organizations.stubs(:fetch_with_app).returns(partners_api_response)
+        Resources::EnvFile.expects(:parse_external_env).returns({ host: "host" })
+        ShopifyCli::PartnersAPI::Organizations.expects(:fetch_with_app).with(@context).returns(partners_api_response)
+        expect_user_prompts
+        env_file = Minitest::Mock.new
+        Resources::EnvFile.expects(:new).with(new_env_file_values.merge({ host: "host" })).returns(env_file)
+        env_file.expect(:write, nil, [@context])
+        assert_equal partners_api_response[1], EnsureEnv.call(@context)
+      end
+
+      def test_not_all_required_values_found
+        Resources::EnvFile.expects(:parse_external_env).returns(existing_env_file_values)
+        ShopifyCli::PartnersAPI::Organizations.expects(:fetch_with_app).with(@context).returns(partners_api_response)
+        expect_user_prompts
+        env_file = Minitest::Mock.new
+        Resources::EnvFile.expects(:new).with(new_env_file_values).returns(env_file)
+        env_file.expect(:write, nil, [@context])
+        assert_equal partners_api_response[1], EnsureEnv.call(@context, required: [:api_key, :secret, :shop])
+      end
+
+      def test_all_required_values_found
+        Resources::EnvFile.expects(:parse_external_env).returns(existing_env_file_values)
+        Resources::EnvFile.expects(:new).never
+        assert_empty(EnsureEnv.call(@context))
+      end
+
+      def test_regenerate_existing_env_file
+        Resources::EnvFile
+          .expects(:parse_external_env)
+          .returns(existing_env_file_values.merge({ shop: "shop", host: "host" }))
+          ShopifyCli::PartnersAPI::Organizations.expects(:fetch_with_app).with(@context).returns(partners_api_response)
+        expect_user_prompts
+        env_file = Minitest::Mock.new
+        Resources::EnvFile.expects(:new).with(new_env_file_values.merge({ host: "host" })).returns(env_file)
+        env_file.expect(:write, nil, [@context])
+        assert_equal partners_api_response[1], EnsureEnv.call(@context, regenerate: true)
+      end
+
+      def test_regenerate_empty_env_file
+        Resources::EnvFile.expects(:parse_external_env).returns({})
+        ShopifyCli::PartnersAPI::Organizations.expects(:fetch_with_app).with(@context).returns(partners_api_response)
+        expect_user_prompts
+        env_file = Minitest::Mock.new
+        Resources::EnvFile.expects(:new).with(new_env_file_values).returns(env_file)
+        env_file.expect(:write, nil, [@context])
+        assert_equal partners_api_response[1], EnsureEnv.call(@context, regenerate: true)
+      end
+
+      private
+
+      def expect_user_prompts
         CLI::UI::Prompt.expects(:ask).with(@context.message('core.tasks.ensure_env.organization_select')).returns(101)
         CLI::UI::Prompt.expects(:ask).with(
           @context.message('core.tasks.ensure_env.development_store_select')
         ).returns('store.myshopify.com')
+      end
 
-        Resources::EnvFile.expects(:new).with(
+      def new_env_file_values
+        {
           api_key: 1235,
           secret: 1234,
           shop: "store.myshopify.com",
           scopes: "write_products,write_customers,write_draft_orders",
-          host: "host",
-          extra: {}
-        ).returns(stub(:write))
-        EnsureEnv.call(@context)
+          extra: {},
+        }
       end
 
-      private
+      def existing_env_file_values
+        { api_key: "apikey", secret: "secret" }
+      end
 
       def partners_api_response
         [{